pyproject.toml
[tool.poetry]
name = "luxtronik"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
websocket-client = "^1.7.0"

[tool.poetry.dev-dependencies]
websockets = "^12.0"
flake8 = "^7.0.0"
pytest = "^8.1.1"
pyinstaller = "^6.5.0"

[tool.poetry.group.dev.dependencies]
setuptools = "^69.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
luxtronik = "Luxtronik.luxtronik:main"
